Create getImi() function
To create this function, I also created an escapeCharacters() function

The jukugo from getYomiJukugo() returns as Japanese characters, but Japanese characters
need to be turned into escaped characters to be used within the Goo dictionary.
In this function, I use Axios to extract the web content and JSDOM to interact with it.
The definition often returns with empty space, so I also trim it before returning it.